% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI_all.R
\name{BMI_all}
\alias{BMI_all}
\title{BMI calculation with DB form that we used}
\usage{
BMI_all()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate csv file that DB form
This fucntion operate without parameter, you just operate function
}
\examples{
BMI_all()

}
