% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI_only.R
\name{BMI_only}
\alias{BMI_only}
\title{BMI calculation}
\usage{
BMI_only()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate BMI and rank of each site
This fucntion operate without parameter, you just operate function
}
\examples{
BMI_only()
}
