% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAI_all.R
\name{FAI_all}
\alias{FAI_all}
\title{FAI calculation}
\usage{
FAI_all(FAI_WS = T, FAI_GR = T)
}
\arguments{
\item{filepath}{Decide whether to set filepath or not}
}
\description{
This function generate FAI and rank of each site and DB format using all survey data
This fucntion operate without parameter, you just operate function
}
\examples{
FAI_all()
}
