% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAI_GR.R
\name{FAI_GR}
\alias{FAI_GR}
\title{FAI calculation by waterside survey result}
\usage{
FAI_GR(filepath = T)
}
\arguments{
\item{filepath}{Decide whether to set filepath or not}
}
\description{
This is a function that generates site-specific FAI, ranking, and DB format using data surveyed at the Greatriver site where center surveying is conducted.
This function operate without parameter, you just operate function
}
\examples{
FAI_GR()
}
