% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAI_GR_only.R
\name{FAI_GR_only}
\alias{FAI_GR_only}
\title{FAI calculation}
\usage{
FAI_GR_only(filepath = T)
}
\arguments{
\item{filepath}{Decide whether to set filepath or not}
}
\description{
This function generate FAI and rank of each site by using greatriver site data that was surveyed in center
This function operate without parameter, you just operate function
}
\examples{
FAI_GR_only()
}
