% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAI_only.R
\name{FAI_only}
\alias{FAI_only}
\title{FAI calculation}
\usage{
FAI_only()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate FAI and rank of each site
This fucntion operate without parameter, you just operate function
}
\examples{
FAI_only()
}
