% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAI_WS.R
\name{FAI_WS}
\alias{FAI_WS}
\title{FAI calculation by waterside survey result}
\usage{
FAI_WS(filepath = T)
}
\arguments{
\item{filepath}{Decide whether to set filepath or not}
}
\description{
This function creates site-specific FAI, ranking, and DB format using data surveyed at the waterside.
This function operate without parameter, you just operate function
}
\examples{
FAI_WS()
}
