% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAI_WS_only.R
\name{FAI_WS_only}
\alias{FAI_WS_only}
\title{FAI calculation}
\usage{
FAI_WS_only(filepath = T)
}
\arguments{
\item{filepath}{Decide whether to set filepath or not}
}
\description{
This function generate FAI and rank from data of waterside
This fucntion operate without parameter, you just operate function
}
\examples{
FAI_WS_only()
}
