% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVI_all.R
\name{RVI_all}
\alias{RVI_all}
\title{RVI calculation with DB form that we used}
\usage{
RVI_all()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate csv file that DB form
This fucntion operate without parameter, you just operate function
}
\examples{
RVI_all()
}
