% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVI_data.transform.R
\name{RVI_data.transform}
\alias{RVI_data.transform}
\title{transfrom the data form for RVI calculation}
\usage{
RVI_data.transform()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate variables for RVI calculation
This fucntion operate without parameter, you just operate function
}
\examples{
RVI_data.transform()
}
