% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVI_only.R
\name{RVI_only}
\alias{RVI_only}
\title{BMI calculation}
\usage{
RVI_only()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate RVI and rank of each site
This fucntion operate without parameter, you just operate function
}
\examples{
RVI_only()
}
