% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDI_all.R
\name{TDI_all}
\alias{TDI_all}
\title{TDI calculation with DB form that we used}
\usage{
TDI_all()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate csv file that DB form
This fucntion operate without parameter, you just operate function
}
\examples{
TDI_all()
}
