% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDI_only.R
\name{TDI_only}
\alias{TDI_only}
\title{TDI calculation}
\usage{
TDI_only()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate BMI and rank of each site
This fucntion operate without parameter, you just operate function
}
\examples{
TDI_only()
}
