% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAI_all.R
\name{FAI_all}
\alias{FAI_all}
\title{FAI calculation with DB form that we used}
\usage{
FAI_all()
}
\arguments{
\item{x}{: Excel file that selected in pop-up window}
}
\description{
This function generate csv file that DB form
This fucntion operate without parameter, you just operate function
}
\examples{
FAI_all()
}
