% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAI_only.R
\name{FAI_only}
\alias{FAI_only}
\title{FAI calculation}
\usage{
FAI_only(FAI_WS_only = T, FAI_GR_only = T)
}
\arguments{
\item{filepath}{Decide whether to set filepath or not}
}
\description{
This function generate FAI and rank of each site using all survey data
This fucntion operate without parameter, you just operate function
}
\examples{
FAI_only()
}
